Remove doubled functions
