+++ app/src/routes/+error.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
     import ChatDisplay from "$components/ChatDisplay.svelte";
-    import { getBadges } from "$lib/preview";
-    import { badges, emotes, isMobile } from "$stores/global";
+    import { getBadges, sendFakeMessage } from "$lib/preview";
+    import { emotes, isMobile } from "$stores/global";
     import { onDestroy, onMount } from "svelte";
 
     import SevenTV_main from "$lib/services/7TV/main";
@@ -61,7 +61,11 @@
             },
         }));
 
-        sendInterval = setInterval(sendFakeMessage, 1000);
+        sendInterval = setInterval(() => {
+            const msg = msgs[Math.floor(Math.random() * msgs.length)];
+
+            sendFakeMessage(msg);
+        }, 1000);
     });
 
     onDestroy(() => clearInterval(sendInterval));
@@ -122,6 +126,18 @@
             position: relative !important;
         }
 
+        #red-dot {
+            position: absolute;
+            background-color: rgb(255, 0, 0);
+            aspect-ratio: 1/1;
+            width: 1rem;
+            border-radius: 1rem;
+            right: -0.25rem;
+            top: -0.25rem;
+
+            animation: pulse 1s ease-in-out infinite;
+        }
+
         user-select: none;
 
         #error {
@@ -139,7 +155,6 @@
 
         min-height: 0;
         border-radius: 1rem;
-        overflow: hidden;
         background-color: #5e5e5e0e;
 
         transform: perspective(1200px) rotateX(var(--chat-x))
@@ -148,6 +163,16 @@
             0 1px 2px rgba(255, 255, 255, 0.2),
             0 4px 8px rgba(0, 0, 0, 0.15),
             0 12px 24px rgba(0, 0, 0, 0.1);
+    }
+
+    @keyframes pulse {
+        0%,
+        100% {
+            opacity: 1;
+        }
+        50% {
+            opacity: 0.5;
+        }
     }
 
     #buttons {
